CSS Cleanup and File Organization
This commit moves the CSS and images to folders to help clean up the
root directory. It also deletes all the previous CSS, none of which was
being applied. Instead, there's now a Sass file that loads normalize
and sets some very basic styling. Finally, the 404 image was set as a
full-screen background image for a nicer look and feel.